Breaching saw
As per request, breaching saw capable of taking down normal walls for now. It's robust as all shit, there's going to be a few "balance" changes to it in the near future. As well how it handles walls will change, and an in-hand when the spriters get around to it. Only being able to cut when active will come into play once I get to slowly grinding through walls.